[add] : 컴포넌트 구조 수정 & router 모듈화
