+++ front/.eslintrc.js
@@ -1,44 +1,48 @@
 module.exports = {
-	root: true,
-	env: {
-		node: true,
-	},
-	extends: [
-		'plugin:vue/essential',
-		'eslint:recommended',
-		'plugin:prettier/recommended',
-	],
-	parserOptions: {
-		parser: '@babel/eslint-parser',
-	},
-	rules: {
-		'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-		'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-		'prettier/prettier': [
-			'error',
-			{
-				singleQuote: true,
-				semi: true,
-				useTabs: true,
-				tabWidth: 2,
-				trailingComma: 'all',
-				printWidth: 80,
-				bracketSpacing: true,
-				arrowParens: 'avoid',
-				endOfLine: 'auto',
-			},
-		],
-		'no-unused-vars': 'off',
-	},
-	overrides: [
-		{
-			files: [
-				'**/__tests__/*.{j,t}s?(x)',
-				'**/tests/unit/**/*.spec.{j,t}s?(x)',
-			],
-			env: {
-				jest: true,
-			},
-		},
-	],
-};
+  root: true,
+  env: {
+    node: true,
+  },
+  extends: [
+    'plugin:vue/essential',
+    'eslint:recommended',
+    'plugin:prettier/recommended',
+  ],
+  parserOptions: {
+    parser: '@babel/eslint-parser',
+  },
+  rules: {
+    'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
+    'prettier/prettier': [
+      'error',
+      {
+        singleQuote: true,
+        semi: false,
+        tabWidth: 2,
+        trailingComma: 'all',
+        printWidth: 80,
+        bracketSpacing: true,
+        arrowParens: 'avoid',
+        endOfLine: 'auto',
+      },
+    ],
+    'no-unused-vars': 'off',
+    'vue/no-unused-components':
+      process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-useless-escape':
+      process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'vue/multi-word-component-names': 'off',
+  },
+  overrides: [
+    {
+      files: [
+        '**/__tests__/*.{j,t}s?(x)',
+        '**/tests/unit/**/*.spec.{j,t}s?(x)',
+      ],
+      env: {
+        jest: true,
+      },
+    },
+  ],
+}
