+++ front/src/components/account/scss/style.scss
