+++ front/src/layouts/Contents.vue
@@ -1,0 +1,9 @@
+<template>
+  <router-view />
+</template>
+
+<script>
+export default {
+  name: 'Contents',
+}
+</script>
