+++ front/src/main.js
@@ -1,10 +1,10 @@
-import Vue from 'vue';
-import App from './App.vue';
-import router from '@/routes/index';
+import Vue from 'vue'
+import App from './App.vue'
+import router from '@/routes/index'
 
-Vue.config.productionTip = false;
+Vue.config.productionTip = false
 
 new Vue({
-	render: h => h(App),
-	router,
-}).$mount('#app');
+  render: h => h(App),
+  router,
+}).$mount('#app')
