--- front/src/views/LoginPage.vue
@@ -1,13 +1,0 @@
-<template>
-	<div>
-		<h1 class="screen_out">로그인 페이지</h1>
-		<LoginForm></LoginForm>
-	</div>
-</template>
-
-<script>
-import LoginForm from '../components/auth/LoginForm.vue';
-export default { components: { LoginForm } };
-</script>
-
-<style></style>
