--- front/src/views/NotFoundPage.vue
@@ -1,9 +1,0 @@
-<template>
-	<div>Page is not found</div>
-</template>
-
-<script>
-export default {};
-</script>
-
-<style></style>
