--- front/src/views/SignupPage.vue
@@ -1,13 +1,0 @@
-<template>
-	<div>
-		<h1 class="screen_out">회원가입 페이지</h1>
-		<SignupForm></SignupForm>
-	</div>
-</template>
-
-<script>
-import SignupForm from '../components/auth/SignupForm.vue';
-export default { components: { SignupForm } };
-</script>
-
-<style></style>
