+++ front/src/views/account/LoginPage.vue
@@ -1,0 +1,18 @@
+<template>
+  <div>
+    <h1 class="screen_out">로그인 페이지</h1>
+    <LoginForm></LoginForm>
+  </div>
+</template>
+
+<script>
+import LoginForm from '@/components/account/LoginForm.vue'
+
+export default {
+  name: 'Login',
+  components: { LoginForm },
+  data() {
+    return {}
+  },
+}
+</script>
