+++ front/src/views/account/SignupPage.vue
@@ -1,0 +1,11 @@
+<template>
+  <div>
+    <h1 class="screen_out">회원가입 페이지</h1>
+    <SignupForm></SignupForm>
+  </div>
+</template>
+
+<script>
+import SignupForm from '@/components/account/SignupForm.vue'
+export default { components: { SignupForm } }
+</script>
