+++ front/src/views/error/NotFoundPage.vue
@@ -1,0 +1,12 @@
+<template>
+  <div>Page is not found</div>
+</template>
+
+<script>
+export default {
+  name: 'NotFound',
+  data() {
+    return {}
+  },
+}
+</script>
