+++ front/src/views/main/scss/main.scss
@@ -1,6 +1,6 @@
 @import '@/styles/variables/index.scss';
 
-.container {
+.main-contents {
 	width: 100%;
 	height: 80vh;
 	color: #fff;
@@ -9,7 +9,7 @@
 	position: relative;
 	z-index: 1;
 }
-.container::after {
+.main-contents::after {
 	background-image: url('@/assets/images/BackgroundImage.png');
 	width: 100%;
 	height: 100%;
@@ -52,7 +52,7 @@ button {
 	font-weight: 400;
 	color: $gray-100;
 }
-.inner-container {
+.inner-contents {
 	display: flex;
 	flex-direction: column;
 	align-items: center;
